requirements.txt
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
numpy==2.3.1
pandas==2.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
openai python-dotenv
pandas
numpy
streamlit
scikit-learn
faker
google-api-python-client
openai
pyspark
matplotlib
wordcloud
joblib